% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-nl.R
\name{hyph.support.nl}
\alias{hyph.support.nl}
\title{Language support for Dutch}
\usage{
hyph.support.nl()
}
\description{
This function adds support for Dutch to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.nl}}).
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
